feat: structured the ui of the game using css
I added some css to the code in order to structure the ui and make it look.
The goal wasnt to make it look good but instead be a draft of what the final
result would look like. The javascript/logic of how the game works is
already done. The only thing left is the css. This commit contains the
basic css.