Done with the project
I implemented the last css and refined the layout and design
of the game to make it look better. It is not refined to look really
good because i didnt want to spend a lot of time on the project.
For example, I skipped having any media querys so my website is not
responsive.